+++ backend/src/Services/Inventory/Domain/Entities/InventoryMovement.cs
@@ -7,7 +7,7 @@ public class InventoryMovement : BaseEntity
 {
     public string TenantId { get; private set; } = string.Empty;
     public Guid StoreInventoryId { get; private set; }
-    public int Quantity { get; private set; }
+    public decimal Quantity { get; private set; }
     public InventoryMovementType Type { get; private set; }
     public string Reason { get; private set; } = string.Empty;
     public Guid UserId { get; private set; }
@@ -15,14 +15,14 @@ public class InventoryMovement : BaseEntity
 
     public Guid? UOMId { get; private set; }
     public string? UOMCode { get; private set; }
-    public int? OriginalQuantity { get; private set; }
-    public int? ConversionFactor { get; private set; }
+    public decimal? OriginalQuantity { get; private set; }
+    public decimal? ConversionFactor { get; private set; }
 
     public StoreInventory StoreInventory { get; private set; } = null!;
 
     private InventoryMovement() { }
 
-    public InventoryMovement(string tenantId, Guid storeInventoryId, int quantity, InventoryMovementType type,
+    public InventoryMovement(string tenantId, Guid storeInventoryId, decimal quantity, InventoryMovementType type,
                            string reason, Guid userId)
     {
         TenantId = tenantId;
@@ -34,8 +34,8 @@ public class InventoryMovement : BaseEntity
         MovementDate = DateTime.UtcNow;
     }
 
-    public InventoryMovement(string tenantId, Guid storeInventoryId, int quantity, InventoryMovementType type,
-                           string reason, Guid userId, Guid? uomId, string? uomCode, int? originalQuantity, int? conversionFactor)
+    public InventoryMovement(string tenantId, Guid storeInventoryId, decimal quantity, InventoryMovementType type,
+                           string reason, Guid userId, Guid? uomId, string? uomCode, decimal? originalQuantity, decimal? conversionFactor)
         : this(tenantId, storeInventoryId, quantity, type, reason, userId)
     {
         UOMId = uomId;
